+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/ContentManagerPlugin.java
@@ -16,53 +16,57 @@
  */
 package com.wazuh.contentmanager;
 
+import java.time.Instant;
+import java.time.temporal.ChronoUnit;
+import java.util.*;
+import java.util.concurrent.Semaphore;
+import java.util.function.Supplier;
+
 import com.wazuh.contentmanager.cti.console.CtiConsole;
 import com.wazuh.contentmanager.index.ConsumersIndex;
 import com.wazuh.contentmanager.index.ContentIndex;
+import com.wazuh.contentmanager.jobscheduler.ContentJobParameter;
+import com.wazuh.contentmanager.jobscheduler.ContentJobRunner;
+import com.wazuh.contentmanager.jobscheduler.jobs.HelloWorldJob;
+import com.wazuh.contentmanager.rest.services.RestDeleteSubscriptionAction;
+import com.wazuh.contentmanager.rest.services.RestGetSubscriptionAction;
+import com.wazuh.contentmanager.rest.services.RestPostSubscriptionAction;
+import com.wazuh.contentmanager.rest.services.RestPostUpdateAction;
 import com.wazuh.contentmanager.settings.PluginSettings;
 import com.wazuh.contentmanager.utils.Privileged;
 import com.wazuh.contentmanager.utils.SnapshotManager;
+
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+
+import org.opensearch.action.index.IndexRequest;
 import org.opensearch.cluster.metadata.IndexNameExpressionResolver;
 import org.opensearch.cluster.node.DiscoveryNode;
 import org.opensearch.cluster.service.ClusterService;
-import org.opensearch.common.settings.Setting;
+import org.opensearch.common.settings.*;
+import org.opensearch.common.xcontent.XContentFactory;
 import org.opensearch.core.common.io.stream.NamedWriteableRegistry;
 import org.opensearch.core.xcontent.NamedXContentRegistry;
 import org.opensearch.env.Environment;
 import org.opensearch.env.NodeEnvironment;
+import org.opensearch.jobscheduler.spi.JobSchedulerExtension;
+import org.opensearch.jobscheduler.spi.ScheduledJobParser;
+import org.opensearch.jobscheduler.spi.ScheduledJobRunner;
+import org.opensearch.jobscheduler.spi.schedule.IntervalSchedule;
+import org.opensearch.plugins.ActionPlugin;
 import org.opensearch.plugins.ClusterPlugin;
 import org.opensearch.plugins.Plugin;
 import org.opensearch.repositories.RepositoriesService;
+import org.opensearch.rest.RestHandler;
 import org.opensearch.script.ScriptService;
 import org.opensearch.threadpool.ThreadPool;
 import org.opensearch.transport.client.Client;
 import org.opensearch.watcher.ResourceWatcherService;
 
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
-import java.util.concurrent.Semaphore;
-import java.util.function.Supplier;
-
-import com.wazuh.contentmanager.jobscheduler.ContentJobParameter;
-import com.wazuh.contentmanager.jobscheduler.ContentJobRunner;
-import com.wazuh.contentmanager.jobscheduler.jobs.HelloWorldJob;
-import org.opensearch.jobscheduler.spi.JobSchedulerExtension;
-import org.opensearch.jobscheduler.spi.ScheduledJobParser;
-import org.opensearch.jobscheduler.spi.ScheduledJobRunner;
-import org.opensearch.jobscheduler.spi.schedule.IntervalSchedule;
-import org.opensearch.action.index.IndexRequest;
-import org.opensearch.common.xcontent.XContentFactory;
-import java.time.Instant;
-import java.time.temporal.ChronoUnit;
-
 /**
  * Main class of the Content Manager Plugin
  */
-public class ContentManagerPlugin extends Plugin implements ClusterPlugin, JobSchedulerExtension {
+public class ContentManagerPlugin extends Plugin implements ClusterPlugin, JobSchedulerExtension, ActionPlugin {
     private static final Logger log = LogManager.getLogger(ContentManagerPlugin.class);
     private static final String JOB_INDEX_NAME = ".wazuh-content-manager-jobs";
 
@@ -70,7 +74,6 @@ public class ContentManagerPlugin extends Plugin implements ClusterPlugin, JobSc
      * Semaphore to ensure the context index creation is only triggered once.
      */
     private static final Semaphore indexCreationSemaphore = new Semaphore(1);
-
     private ConsumersIndex consumersIndex;
     private ContentIndex contentIndex;
     private SnapshotManager snapshotManager;
@@ -78,6 +81,11 @@ public class ContentManagerPlugin extends Plugin implements ClusterPlugin, JobSc
     private ClusterService clusterService;
     private CtiConsole ctiConsole;
     private Client client;
+
+    // Rest API endpoints
+    public static final String PLUGINS_BASE_URI = "/_plugins/content-manager";
+    public static final String SUBSCRIPTION_URI = PLUGINS_BASE_URI + "/subscription";
+    public static final String UPDATE_URI = PLUGINS_BASE_URI + "/update";
 
     @Override
     public Collection<Object> createComponents(
@@ -100,9 +108,11 @@ public class ContentManagerPlugin extends Plugin implements ClusterPlugin, JobSc
         this.contentIndex = new ContentIndex(client);
         this.snapshotManager =
                 new SnapshotManager(environment, this.consumersIndex, this.contentIndex, new Privileged());
-//        this.ctiConsole = new CtiConsole(new AuthServiceImpl());
         ContentJobRunner runner = ContentJobRunner.getInstance();
         runner.registerExecutor(HelloWorldJob.JOB_TYPE, new HelloWorldJob());
+            new SnapshotManager(environment, this.consumersIndex, this.contentIndex, new Privileged());
+        // Content Manager 5.0
+        this.ctiConsole = new CtiConsole();
         return Collections.emptyList();
     }
 
@@ -117,7 +127,6 @@ public class ContentManagerPlugin extends Plugin implements ClusterPlugin, JobSc
     public void onNodeStarted(DiscoveryNode localNode) {
         // Only cluster managers are responsible for the initialization.
         if (localNode.isClusterManagerNode()) {
-            log.info("Starting Content Manager plugin initialization");
             this.start();
         }
         this.scheduleHelloWorldJob();
@@ -155,6 +164,22 @@ public class ContentManagerPlugin extends Plugin implements ClusterPlugin, JobSc
             log.info("Resource token {}", resourceToken);
         }
         */
+    }
+
+    public List<RestHandler> getRestHandlers(
+        Settings settings,
+        org.opensearch.rest.RestController restController,
+        org.opensearch.common.settings.ClusterSettings clusterSettings,
+        org.opensearch.common.settings.IndexScopedSettings indexScopedSettings,
+        org.opensearch.common.settings.SettingsFilter settingsFilter,
+        org.opensearch.cluster.metadata.IndexNameExpressionResolver indexNameExpressionResolver,
+        java.util.function.Supplier<org.opensearch.cluster.node.DiscoveryNodes> nodesInCluster) {
+        return List.of(
+            new RestGetSubscriptionAction(this.ctiConsole),
+            new RestPostSubscriptionAction(this.ctiConsole),
+            new RestDeleteSubscriptionAction(this.ctiConsole),
+            new RestPostUpdateAction(this.ctiConsole)
+        );
     }
 
     /**