+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/client/CTIClient.java
@@ -27,10 +27,6 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.opensearch.env.Environment;
 
-import org.opensearch.action.ActionType;
-import org.opensearch.action.ActionRequest;
-import org.opensearch.core.action.ActionResponse;
-
 import java.io.*;
 import java.net.URI;
 import java.net.URISyntaxException;
@@ -41,8 +37,8 @@ import java.time.Duration;
 import java.time.ZonedDateTime;
 import java.util.*;
 
-import com.wazuh.contentmanager.model.cti.Changes;
-import com.wazuh.contentmanager.model.cti.ConsumerInfo;
+import com.wazuh.contentmanager.cti.catalog.model.Changes;
+import com.wazuh.contentmanager.cti.catalog.model.ConsumerInfo;
 import com.wazuh.contentmanager.settings.PluginSettings;
 import com.wazuh.contentmanager.utils.VisibleForTesting;
 import com.wazuh.contentmanager.utils.XContentUtils;