+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/cti/catalog/model/Offset.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.wazuh.contentmanager.model.cti;
+package com.wazuh.contentmanager.cti.catalog.model;
 
 import org.opensearch.core.xcontent.ToXContentObject;
 import org.opensearch.core.xcontent.XContentBuilder;